+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Adds an error modal to the bid flow - sweir27
 * The country select form should pre-populate if user have already selected a country - yuki2
 * Update graphql + use new argument to fetch increments - sweir27
+* Sets artwork and artist name labels on ConfirmBid component to use only one line, and truncate the tail - ash
 
 ### 1.5.10